Move profile related files from seminar to user package
