+++ seminar2/src/main/kotlin/com/wafflestudio/seminar/core/user/database/InstructorProfileEntity.kt
@@ -1,15 +1,14 @@
-package com.wafflestudio.seminar.core.seminar.database
+package com.wafflestudio.seminar.core.user.database
 
 import com.wafflestudio.seminar.common.BaseTimeEntity
-import com.wafflestudio.seminar.core.user.database.UserEntity
 import javax.persistence.*
 
 @Entity
-@Table(name = "participantProfile")
-class ParticipantProfileEntity(
+@Table(name = "instructorProfile")
+class InstructorProfileEntity(
     @OneToOne(cascade = [CascadeType.ALL])
     @JoinColumn(name = "user_id", referencedColumnName = "id")
     val user: UserEntity,
-    val university: String,
-    val isRegistered: Boolean,
+    val company: String,
+    val year: Int?,
 ) : BaseTimeEntity()
