+++ seminar2/src/main/kotlin/com/wafflestudio/seminar/core/user/database/ParticipantProfileEntity.kt
@@ -1,15 +1,14 @@
-package com.wafflestudio.seminar.core.seminar.database
+package com.wafflestudio.seminar.core.user.database
 
 import com.wafflestudio.seminar.common.BaseTimeEntity
-import com.wafflestudio.seminar.core.user.database.UserEntity
 import javax.persistence.*
 
 @Entity
-@Table(name = "instructorProfile")
-class InstructorProfileEntity(
+@Table(name = "participantProfile")
+class ParticipantProfileEntity(
     @OneToOne(cascade = [CascadeType.ALL])
     @JoinColumn(name = "user_id", referencedColumnName = "id")
     val user: UserEntity,
-    val company: String,
-    val year: Int?,
+    val university: String,
+    val isRegistered: Boolean,
 ) : BaseTimeEntity()
