+++ seminar2/src/main/kotlin/com/wafflestudio/seminar/core/user/domain/InstructorProfile.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.seminar.core.seminar.domain
+package com.wafflestudio.seminar.core.user.domain
 
 import com.wafflestudio.seminar.core.user.database.UserEntity
